+++ spec/flows/flow/router/simple_spec.rb
@@ -3,7 +3,7 @@ require 'spec_helper'
 RSpec.describe Flows::Flow::Router::Simple do
   include Flows::Result::Helpers
 
-  describe '.call' do
+  describe '#call' do
     subject(:router) { described_class.new(:success, :failure) }
 
     it 'returns success path for success result' do
@@ -13,5 +13,13 @@ RSpec.describe Flows::Flow::Router::Simple do
     it 'returns failure path for success result' do
       expect(router.call(err)).to eq :failure
     end
+  end
+
+  describe '#destinations' do
+    subject(:destinations) { router.destinations }
+
+    let(:router) { described_class.new(:a, :b) }
+
+    it { is_expected.to eq %i[a b] }
   end
 end